fix: expand error handling for github api interactions
GitHub API interaction error handling was setup to exit the app on error. This was causing the pods to restart on "server misbehaving" and other issues not caused by the app itself. Error handling is expanded on by allowing 5 logged errors before exiting. Further error handling will be built as a better knowledge base of common issues can be defined and handled.